Break up test code
