+++ lib/producer.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const assert = require('assert');
 const nsq = require('nsqjs');
 const nsqlookup = require('nsq-lookup');
 const once = require('once');
